refactor: fix a misspelling, change criptography to cryptography
