+++ src/data/test/index.ts
@@ -1,4 +1,4 @@
-export * from './mock-criptography'
+export * from './mock-cryptography'
 export * from './mock-db-account'
 export * from './mock-db-log'
 export * from './mock-db-survey'
