+++ src/data/usecases/account/add-account/db-add-account-protocols.ts
@@ -1,5 +1,5 @@
 export * from '@/domain/usecases/account/add-account'
 export * from '@/domain/models/account'
-export * from '@/data/protocols/criptography/hasher'
+export * from '@/data/protocols/cryptography/hasher'
 export * from '@/data/protocols/db/account/add-account-repository'
 export * from '@/data/protocols/db/account/load-account-by-email-repository'
