+++ src/data/usecases/account/authentication/db-authentication-protocols.ts
@@ -3,5 +3,5 @@ export * from '@/domain/models/authentication'
 export * from '@/domain/usecases/account/authentication'
 export * from '@/data/protocols/db/account/load-account-by-email-repository'
 export * from '@/data/protocols/db/account/update-access-token-repository'
-export * from '@/data/protocols/criptography/hash-comparer'
-export * from '@/data/protocols/criptography/encrypter'
+export * from '@/data/protocols/cryptography/hash-comparer'
+export * from '@/data/protocols/cryptography/encrypter'
