+++ src/data/usecases/account/load-account-by-token/db-load-account-by-token-protocols.ts
@@ -1,4 +1,4 @@
-export * from '@/data/protocols/criptography/decrypter'
+export * from '@/data/protocols/cryptography/decrypter'
 export * from '@/data/protocols/db/account/load-account-by-token-repository'
 export * from '@/domain/usecases/account/load-account-by-token'
 export * from '@/domain/models/account'
