+++ src/main/factories/usecases/account/add-account/db-add-account-factory.ts
@@ -1,7 +1,7 @@
 import { DbAddAccount } from '@/data/usecases/account/add-account/db-add-account'
 import { AddAccount } from '@/domain/usecases/account/add-account'
 import { AccountMongoRepository } from '@/infra/db/mongodb/account/account-mongo-repository'
-import { BcryptAdapter } from '@/infra/criptography/bcrypt-adapter/bcrypt-adapter'
+import { BcryptAdapter } from '@/infra/cryptography/bcrypt-adapter/bcrypt-adapter'
 
 export const makeDbAddAccount = (): AddAccount => {
   const salt = 12
