+++ src/main/factories/usecases/account/authentication/db-authentication-factory.ts
@@ -1,7 +1,7 @@
 import env from '@/main/config/env'
 import { AccountMongoRepository } from '@/infra/db/mongodb/account/account-mongo-repository'
-import { BcryptAdapter } from '@/infra/criptography/bcrypt-adapter/bcrypt-adapter'
-import { JwtAdapter } from '@/infra/criptography/jwt-adapter/jwt-adapter'
+import { BcryptAdapter } from '@/infra/cryptography/bcrypt-adapter/bcrypt-adapter'
+import { JwtAdapter } from '@/infra/cryptography/jwt-adapter/jwt-adapter'
 import { DbAuthentication } from '@/data/usecases/account/authentication/db-authentication'
 import { Authentication } from '@/domain/usecases/account/authentication'
