+++ src/main/factories/usecases/account/load-account-by-token/db-load-account-by-token-factory.ts
@@ -2,7 +2,7 @@ import env from '@/main/config/env'
 import { LoadAccountByToken } from '@/domain/usecases/account/load-account-by-token'
 import { DbLoadAccountByToken } from '@/data/usecases/account/load-account-by-token/db-load-account-by-token'
 import { AccountMongoRepository } from '@/infra/db/mongodb/account/account-mongo-repository'
-import { JwtAdapter } from '@/infra/criptography/jwt-adapter/jwt-adapter'
+import { JwtAdapter } from '@/infra/cryptography/jwt-adapter/jwt-adapter'
 
 export const makeDbLoadAccountByToken = (): LoadAccountByToken => {
   const jwtAdapter = new JwtAdapter(env.jwtSecret)
